pyproject.toml
[project]
name = "fastapi-with-otel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "pydantic>=2.11.7",
    "uvicorn>=0.34.3",
]
